Add incremental progress reporting system.
No low-level code has been implemented with this tooling yet, but once
it is, high-level code will be able to produce progress bars and
possibly other displays of how long various slow operations are taking.

This does include an option to turn on progress via a Click-based
progress bar in the butler command, but progress displays are off by
default even there.